refactor: replace runtime registration with AST-based analysis
Remove Module.register() and the runtime reflection-based _typedefs()
implementation. Both registration paths (CLI --register and serve)
now delegate to a shared ast_register() pipeline that uses static
AST analysis instead of importing the module.

Extract source file discovery functions into _discovery.py to avoid
circular imports and provide a clean dependency graph. Also fix a
pre-existing bug where single-file modules were unreachable due to
an early return in find_source_files().
